Malam Yang Panjang

+++ routes/web.php
@@ -72,16 +72,6 @@ Route::group(['middleware' => ['auth']], function () {
             Route::delete('/pengaturan/kategori-menu/{id}','delete')->name('pengaturan.kategorimenu.delete');
         });
 
-        // Lokasi
-        Route::controller(LokasiController::class)->group(function()
-        {
-            Route::get('lokasi', 'index')->name('lokasi.index');
-            Route::get('lokasi/create', 'create')->name('lokasi.create');
-            Route::post('lokasi/add', 'lokasi_add')->name('lokasi.add');
-            Route::delete('lokasi/{id}/delete','delete')->name('lokasi.delete');
-            Route::put('lokasi/{id}', 'update')->name('lokasi.update');
-        });
-
         // Menu
         Route::controller(MenuController::class)->group(function(){
             Route::get('/pengaturan/menu','index')->name('pengaturan.menu.index');
@@ -89,14 +79,6 @@ Route::group(['middleware' => ['auth']], function () {
             Route::put('/pengaturan/menu/{id}','update')->name('pengaturan.menu.update');
             Route::put('/pengaturan/menu/{id}/show','updateShow')->name('pengaturan.menu.updateShow');
             Route::delete('/pengaturan/menu/{id}','delete')->name('pengaturan.menu.delete');
-        });
-
-        // SubMenu
-        Route::controller(SubMenuController::class)->group(function(){
-            Route::get('/pengaturan/sub-menu','index')->name('pengaturan.submenu.index');
-            Route::post('/pengaturan/sub-menu','store')->name('pengaturan.submenu.store');
-            Route::put('/pengaturan/sub-menu/{id}','update')->name('pengaturan.submenu.update');
-            Route::delete('/pengaturan/sub-menu/{id}','delete')->name('pengaturan.submenu.delete');
         });
 
         // Ajax Menu Request
@@ -107,6 +89,14 @@ Route::group(['middleware' => ['auth']], function () {
             Route::post('/ajax/sub-menu/edit','getSubMenuEdit')->name('ajax.getSubMenuEdit');
             Route::get('/ajax/kategori-menu','getKategoriMenu')->name('ajax.getKategoriMenu');
             Route::post('/ajax/kategori-menu/edit','getKategoriMenuEdit')->name('ajax.getKategoriMenuEdit');
+        });
+
+        // SubMenu
+        Route::controller(SubMenuController::class)->group(function(){
+            Route::get('/pengaturan/sub-menu','index')->name('pengaturan.submenu.index');
+            Route::post('/pengaturan/sub-menu','store')->name('pengaturan.submenu.store');
+            Route::put('/pengaturan/sub-menu/{id}','update')->name('pengaturan.submenu.update');
+            Route::delete('/pengaturan/sub-menu/{id}','delete')->name('pengaturan.submenu.delete');
         });
 
         // Role
@@ -142,26 +132,58 @@ Route::group(['middleware' => ['auth']], function () {
         });
     });
 
-    // Izin
+    /*
+    |--------------------------------------------------------------------------
+    | Middleware Untuk Siapa Saja Yang Memiliki Role & Sudah Login
+    |--------------------------------------------------------------------------
+    */
+
+    /*
+    | Route Dashboard
+    | ----------------------
+    */
+    Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
+
+    /*
+    | Route Profile
+    | ----------------------
+    */
+    Route::controller(ProfilController::class)->group(function(){
+        Route::get('/profil', 'index')->name('profile.index');
+        Route::get('/profil/reset-password', 'indexResetPassword')->name('profile.indexResetPassword');
+        Route::put('/profil/{id}', 'updateProfil')->name('profile.updateProfil');
+        Route::put('/profil/{id}/password', 'updatePassword')->name('profile.updatePassword');
+        Route::put('/profil/{id}/image', 'updateImage')->name('profile.updateImage');
+        Route::put('/profil/{id}/ttd', 'updateTtd')->name('profile.updateTtd');
+    });
+
+    /*
+    | Route Izin
+    | ----------------------
+    */
     Route::controller(IzinController::class)->group(function()
     {
         // Izin untuk tampilan Teknisi
-        Route::get('izin', 'index')->name('izin.index');
-        Route::get('izin/create', 'create')->name('izin.create');
-        Route::post('izin', 'store')->name('izin.store');
-        Route::get('izin/{id}/edit', 'edit')->name('izin.edit');
-        Route::put('izin/{id}', 'update')->name('izin.update');
-        Route::put('izin/{id}/validasi', 'updateValidasi')->name('izin.updateValidasi');
-        Route::delete('izin/{id}', 'delete')->name('izin.delete');
+        Route::get('administrasi/izin', 'index')->name('izin.index')->middleware('permission:izin_read');
+        Route::get('administrasi/izin/create', 'create')->name('izin.create')->middleware('permission:izin_create');
+        Route::post('administrasi/izin', 'store')->name('izin.store')->middleware('permission:izin_create');
+        Route::get('administrasi/izin/{id}/edit', 'edit')->name('izin.edit')->middleware('permission:izin_update');
+        Route::put('administrasi/izin/{id}', 'update')->name('izin.update')->middleware('permission:izin_update');
+        Route::put('administrasi/izin/{id}/validasi', 'updateValidasi')->name('izin.updateValidasi')->middleware('permission:izin_update');
+        Route::delete('administrasi/izin/{id}', 'delete')->name('izin.delete')->middleware('permission:izin_delete');
 
         // Izin untuk admin (setting)
-        Route::get('pengaturan/izin', 'indexPengaturan')->name('pengaturan.izin.index');
-        Route::get('pengaturan/izin/create', 'createPengaturan')->name('pengaturan.izin.create');
-        Route::post('pengaturan/izin', 'storePengaturan')->name('pengaturan.izin.store');
-        Route::put('pengaturan/izin/{id}', 'updatePengaturan')->name('pengaturan.izin.update');
-        Route::delete('pengaturan/izin/{id}','deletePengaturan')->name('pengaturan.izin.delete');
+        Route::get('pengaturan/jumlah-izin', 'indexPengaturan')->name('pengaturan.izin.index')->middleware('permission:pengaturan_izin_read');;
+        Route::get('pengaturan/jumlah-izin/create', 'createPengaturan')->name('pengaturan.izin.create')->middleware('permission:pengaturan_izin_create');;
+        Route::post('pengaturan/jumlah-izin', 'storePengaturan')->name('pengaturan.izin.store')->middleware('permission:pengaturan_izin_create');;
+        Route::put('pengaturan/jumlah-izin/{id}', 'updatePengaturan')->name('pengaturan.izin.update')->middleware('permission:pengaturan_izin_update');;
+        Route::delete('pengaturan/jumlah-izin/{id}','deletePengaturan')->name('pengaturan.izin.delete')->middleware('permission:pengaturan_izin_delete');;
     });
 
+    /*
+    | Route Ajax Izin
+    | ----------------------
+    */
     Route::controller(AjaxIzinController::class)->group(function(){
         Route::get('ajax/izin', 'getIzin')->name('ajax.getIzin');
         Route::post('ajax/izin-valid', 'getValidIzin')->name('ajax.getValidIzin');
@@ -170,7 +192,10 @@ Route::group(['middleware' => ['auth']], function () {
         Route::post('ajax/jumlah-izin/edit', 'getJumlahIzinEdit')->name('ajax.getJumlahIzinEdit');
     });
 
-    // Regional
+    /*
+    | Route Regional
+    | ----------------------
+    */
     Route::controller(RegionalController::class)->group(function()
     {
         Route::get('/pengaturan/regional', 'index')->name('regional.index')->middleware('permission:regional_read');
@@ -181,7 +206,10 @@ Route::group(['middleware' => ['auth']], function () {
         Route::put('/pengaturan/regional/{id}', 'update')->name('regional.update')->middleware('permission:regional_update');
     });
 
-    // Ajax Regional Request
+    /*
+    | Route Ajax Regional
+    | ----------------------
+    */
     Route::controller(AjaxRegionalController::class)->group(function(){
         Route::get('/ajax/regional','getRegional')->name('ajax.getRegional')->middleware('permission:regional_read');
         Route::post('/ajax/regional/map','getRegionalMap')->name('ajax.getRegionalMap')->middleware('permission:regional_read');
@@ -189,32 +217,48 @@ Route::group(['middleware' => ['auth']], function () {
         Route::post('/ajax/regional/edit','getRegionalEdit')->name('ajax.getRegionalEdit')->middleware('permission:user_update');
     });
 
-    // Shift
+    /*
+    | Route Shift
+    | ----------------------
+    */
     Route::resource('/pengaturan/shift',ShiftController::class);
 
-    // Ajax Shift Request
+    /*
+    | Route Ajax Shift
+    | ----------------------
+    */
     Route::controller(AjaxShiftController::class)->group(function(){
         Route::get('/ajax/shift','getShift')->name('ajax.getShift')->middleware('permission:shift_read');
         Route::post('/ajax/shift/edit','getShiftEdit')->name('ajax.getShiftEdit')->middleware('permission:shift_update');
     });
 
-    // Lokasi
+    /*
+    | Route Lokasi
+    | ----------------------
+    */
     Route::controller(LokasiController::class)->group(function()
     {
-        Route::get('lokasi', 'index')->name('lokasi.index');
-        Route::get('lokasi/create', 'create')->name('lokasi.create');
-        Route::post('lokasi/add', 'store')->name('lokasi.add');
-        Route::delete('lokasi/{id}/delete','delete')->name('lokasi.delete');
-        Route::get('lokasi/{id}/edit', 'edit')->name('lokasi.edit');
-        Route::put('lokasi/{id}', 'update')->name('lokasi.update');
+        Route::get('/pengaturan/lokasi', 'index')->name('lokasi.index')->middleware('permission:lokasi_read');
+        Route::get('/pengaturan/lokasi/create', 'create')->name('lokasi.create')->middleware('permission:lokasi_create');
+        Route::post('/pengaturan/lokasi/add', 'store')->name('lokasi.add')->middleware('permission:lokasi_create');
+        Route::delete('/pengaturan/lokasi/{id}/delete','delete')->name('lokasi.delete')->middleware('permission:lokasi_delete');
+        Route::get('/pengaturan/lokasi/{id}/edit', 'edit')->name('lokasi.edit')->middleware('permission:lokasi_update');
+        Route::put('/pengaturan/lokasi/{id}', 'update')->name('lokasi.update')->middleware('permission:lokasi_update');
     });
 
-    // Ajax Lokasi Request
+    /*
+    | Route Ajax Lokasi
+    | ----------------------
+    */
     Route::controller(AjaxLokasiController::class)->group(function(){
-        Route::get('/ajax','getLokasi')->name('ajax.getLokasi');
+        Route::get('/ajax/lokasi','getLokasi')->name('ajax.getLokasi')->middleware('permission:lokasi_read');
+        Route::get('/ajax/lokasi/all-map','getAllLokasiMap')->name('ajax.getAllLokasiMap')->middleware('permission:lokasi_read');
     });
 
-    // User
+    /*
+    | Route User
+    | ----------------------
+    */
     Route::controller(UserController::class)->group(function()
     {
         Route::get('administrasi/user', 'index')->name('user.index')->middleware('permission:user_read');
@@ -226,26 +270,19 @@ Route::group(['middleware' => ['auth']], function () {
         Route::delete('administrasi/user/{id}','delete')->name('user.delete')->middleware('permission:user_delete');
     });
 
-    // Ajax User Request
+    /*
+    | Route Ajax User
+    | ----------------------
+    */
     Route::controller(AjaxUserController::class)->group(function(){
         Route::get('/ajax/user','getUser')->name('ajax.getUser')->middleware('permission:user_read');
         Route::post('/ajax/user/detail','getUserDetail')->name('ajax.getUserDetail')->middleware('permission:user_read');
     });
 
-    // Dashboard
-    Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
-
-    // Profile
-    Route::controller(ProfilController::class)->group(function(){
-        Route::get('/profil', 'index')->name('profile.index');
-        Route::get('/profil/reset-password', 'indexResetPassword')->name('profile.indexResetPassword');
-        Route::put('/profil/{id}', 'updateProfil')->name('profile.updateProfil');
-        Route::put('/profil/{id}/password', 'updatePassword')->name('profile.updatePassword');
-        Route::put('/profil/{id}/image', 'updateImage')->name('profile.updateImage');
-        Route::put('/profil/{id}/ttd', 'updateTtd')->name('profile.updateTtd');
-    });
-
-    // Absen
+    /*
+    | Route Absen
+    | ----------------------
+    */
     Route::controller(AbsenController::class)->group(function(){
         Route::get('/administrasi/absen', 'index')->name('absen.index')->middleware('permission:absen_read');
         Route::get('/administrasi/absen/detail', 'detail')->name('absen.detail')->middleware('permission:absen_read');
@@ -253,7 +290,10 @@ Route::group(['middleware' => ['auth']], function () {
         Route::post('/administrasi/absen', 'store')->name('absen.store')->middleware('permission:absen_create');
     });
 
-    // Ajax Absen Request
+    /*
+    | Route Ajax Absen
+    | ----------------------
+    */
     Route::controller(AjaxAbsenController::class)->group(function(){
         Route::post('/ajax/absen-shift','getAbsenShift')->name('ajax.getAbsenShift');
         Route::get('/ajax/absen-log','getAbsenLog')->name('ajax.getAbsenLog');
